Moved settings#upload method out of settings controller
