+++ core/frontend/services/routing/index.js
@@ -17,5 +17,9 @@ module.exports = {
 
     get TaxonomyRouter() {
         return require('./TaxonomyRouter');
+    },
+
+    get settings() {
+        return require('./SettingsHandler');
     }
 };
