+++ core/server/api/v2/settings.js
@@ -1,11 +1,10 @@
 const Promise = require('bluebird');
 const _ = require('lodash');
-const moment = require('moment-timezone');
 const fs = require('fs-extra');
 const path = require('path');
 const config = require('../../config');
 const models = require('../../models');
-const urlService = require('../../../frontend/services/url');
+const frontendRouting = require('../../../frontend/services/routing');
 const common = require('../../lib/common');
 const settingsCache = require('../../services/settings/cache');
 
@@ -151,62 +150,7 @@ module.exports = {
             method: 'edit'
         },
         query(frame) {
-            const backupRoutesPath = path.join(config.getContentPath('settings'), `routes-${moment().format('YYYY-MM-DD-HH-mm-ss')}.yaml`);
-
-            return fs.copy(`${config.getContentPath('settings')}/routes.yaml`, backupRoutesPath)
-                .then(() => {
-                    return fs.copy(frame.file.path, `${config.getContentPath('settings')}/routes.yaml`);
-                })
-                .then(() => {
-                    urlService.resetGenerators({releaseResourcesOnly: true});
-                })
-                .then(() => {
-                    const siteApp = require('../../web/site/app');
-
-                    const bringBackValidRoutes = () => {
-                        urlService.resetGenerators({releaseResourcesOnly: true});
-
-                        return fs.copy(backupRoutesPath, `${config.getContentPath('settings')}/routes.yaml`)
-                            .then(() => {
-                                return siteApp.reload();
-                            });
-                    };
-
-                    try {
-                        siteApp.reload();
-                    } catch (err) {
-                        return bringBackValidRoutes()
-                            .finally(() => {
-                                throw err;
-                            });
-                    }
-
-                    let tries = 0;
-
-                    function isBlogRunning() {
-                        return Promise.delay(1000)
-                            .then(() => {
-                                if (!urlService.hasFinished()) {
-                                    if (tries > 5) {
-                                        throw new common.errors.InternalServerError({
-                                            message: 'Could not load routes.yaml file.'
-                                        });
-                                    }
-
-                                    tries = tries + 1;
-                                    return isBlogRunning();
-                                }
-                            });
-                    }
-
-                    return isBlogRunning()
-                        .catch((err) => {
-                            return bringBackValidRoutes()
-                                .finally(() => {
-                                    throw err;
-                                });
-                        });
-                });
+            return frontendRouting.settings.activate(frame.file.path);
         }
     },
